Duplicating pages will now duplicate all children too.

+++ system/cms/modules/pages/controllers/admin.php
@@ -148,10 +148,7 @@ class Admin extends Admin_Controller {
 		$order = $this->input->post('order');
 
 		if (is_array($order))
-		{
-			//reset all parent > child relations
-			$this->page_m->update_all(array('parent_id' => 0));
-			
+		{	
 			foreach ($order as $i => $page)
 			{
 				//set the order of the root pages
@@ -206,37 +203,57 @@ class Admin extends Admin_Controller {
 	 * Duplicate a page
 	 * @access public
 	 * @param int $id The ID of the page
+	 * @param int $id The ID of the parent page, if this is a recursive nested duplication
 	 * @return void
 	 */
-	public function duplicate($id = 0)
+	public function duplicate($id, $parent_id = null)
 	{
 		$page  = $this->page_m->get($id);
 		
+		// Steal their children
+		$children = $this->page_m->get_many_by('parent_id', $id);
+		
 		$new_slug = $page->slug;
 		
-		do
+		// No parent around? Do what you like
+		if (is_null($parent_id))
 		{
-			// Turn "foo" into "foo-1"
-			$new_slug = increment_string($new_slug, '-', 2);
-			
-			// Find if this already exists in this level
-			$dupes = $this->page_m->count_by(array(
-				'slug' => $new_slug,
-				'parent_id' => $page->parent_id,
-			));
+			do
+			{
+				// Turn "Foo" into "Foo 2"
+				$page->title = increment_string($page->title, ' ', 2);
+					
+				// Turn "foo" into "foo-2"
+				$page->slug = increment_string($page->slug, '-', 2);
+
+				// Find if this already exists in this level
+				$dupes = $this->page_m->count_by(array(
+					'slug' => $page->slug,
+					'parent_id' => $page->parent_id,
+				));
+			}
+			while ($dupes > 0);
 		}
-		while ($dupes > 0);
 		
-		// Turn "Foo" into "Foo 2"
-		$page->title = increment_string($page->title, ' ', 2);
-		$page->slug = $new_slug;
+		// Oop, a parent turned up, work with that
+		else
+		{
+			$page->parent_id = $parent_id;
+		}
 		
 		$chunks = $this->db->get_where('page_chunks', array('page_id' => $page->id))->result();
 		
-		$id = $this->page_m->insert((array) $page, $chunks);
+		$new_page_id = $this->page_m->insert((array) $page, $chunks);
 		
-		redirect('admin/pages/edit/'.$id);
+		foreach ($children as $child)
+		{
+			$this->duplicate($child->id, $new_page_id);
+		}
 		
+		if ($parent_id === NULL)
+		{
+			redirect('admin/pages/edit/'.$new_page_id);
+		}
 	}
 
 	/**
@@ -530,40 +547,43 @@ class Admin extends Admin_Controller {
 		$ids = ($id) ? array($id) : $this->input->post('action_to');
 
 		// Go through the array of slugs to delete
-		foreach ($ids as $id)
+		if ( ! empty($ids))
 		{
-			if ($id !== 1)
+			foreach ($ids as $id)
 			{
-				$deleted_ids = $this->page_m->delete($id);
+				if ($id !== 1)
+				{
+					$deleted_ids = $this->page_m->delete($id);
 
-				// Wipe cache for this model, the content has changd
-				$this->pyrocache->delete_all('page_m');
-				$this->pyrocache->delete_all('navigation_m');
+					// Wipe cache for this model, the content has changd
+					$this->pyrocache->delete_all('page_m');
+					$this->pyrocache->delete_all('navigation_m');
+				}
+
+				else
+				{
+					$this->session->set_flashdata('error', lang('pages_delete_home_error'));
+				}
+			}
+			
+			// Some pages have been deleted
+			if ( ! empty($deleted_ids))
+			{
+				// Only deleting one page
+				if ( count($deleted_ids) == 1 )
+				{
+					$this->session->set_flashdata('success', sprintf(lang('pages_delete_success'), $deleted_ids[0]));
+				}
+				else // Deleting multiple pages
+				{
+					$this->session->set_flashdata('success', sprintf(lang('pages_mass_delete_success'), count($deleted_ids)));
+				}
 			}
 
-			else
+			else // For some reason, none of them were deleted
 			{
-				$this->session->set_flashdata('error', lang('pages_delete_home_error'));
+				$this->session->set_flashdata('notice', lang('pages_delete_none_notice'));
 			}
-		}
-
-		// Some pages have been deleted
-		if ( ! empty($deleted_ids))
-		{
-			// Only deleting one page
-			if ( count($deleted_ids) == 1 )
-			{
-				$this->session->set_flashdata('success', sprintf(lang('pages_delete_success'), $deleted_ids[0]));
-			}
-			else // Deleting multiple pages
-			{
-				$this->session->set_flashdata('success', sprintf(lang('pages_mass_delete_success'), count($deleted_ids)));
-			}
-		}
-
-		else // For some reason, none of them were deleted
-		{
-			$this->session->set_flashdata('notice', lang('pages_delete_none_notice'));
 		}
 
 		redirect('admin/pages');
